Switch to more modern password storage
rc4 and md5 have been considered obsolete for years, so let's switch to
something more modern: scrypt and aes.

By keeping the old md5/arc password checkers/decoders, we remain
backwards compatible with old configfiles and xml files, but we will
only write new data using the more secure hashing and encryption.

First hashing: pbkdf2. Consensus in the crypto community is that scrypt
is the best key distribution function today, but it (intentionally) is
very memory intensive. We choose pbkdf2 which is still very strong (and
definitely stronger than md5) with parameters suitable for 2016.

Then encryption: AES-GCM. AES is modern, supported by gcrypt, and using
it in counter mode we can even check whether password decryption failed
and fall back to our old arc scheme for backwards compatibility. On
systems where AES-GCM is not available, we keep using the old rc4
encryption.

And finally, update the example configfile to not have old-style hashes.

+++ lib/password.h
@@ -6,6 +6,6 @@ int password_verify(const char *password, const char *hash);
 size_t password_encode(const char *password, char **encoded);
 size_t password_decode(const char *encoded, char **password);
 size_t password_encrypt(const char *password, const char *encryption_key, char **crypt);
-size_t password_decrypt(const char *encrypted, const char *encryption_key, char **password);
+int password_decrypt(const char *encrypted, const char *encryption_key, char **password);
 
 #endif
